Subindo o codigo de autenticacao do usuario

+++ wayne_api/routers/equipment_routers.py
@@ -3,16 +3,20 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 from wayne_api.database import get_session
-from wayne_api.models import Equipment
+from wayne_api.models import Equipment, User
 from wayne_api.schemas import EquipmentBase, EquipmentPublic, EquipmentPartialUpdate, EquipmentList
+from wayne_api.dependencies import verify_token
 
 router = APIRouter(
     prefix="/equipment",
-    tags =["Equipment"]
+    tags =["Equipment"],
+    dependencies=[Depends(verify_token)]
 )
 
 @router.post("/", response_model=EquipmentPublic, status_code=status.HTTP_201_CREATED)
-def create_equipment(equipment: EquipmentBase, session: Session = Depends(get_session)):
+def create_equipment(equipment: EquipmentBase, session: Session = Depends(get_session), current_user: User = Depends(verify_token)):
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Operation not permitted")
     equipment = Equipment(**equipment.model_dump())
     session.add(equipment)
     session.commit()
@@ -36,10 +40,12 @@ def get_equipment(equipment_id: int, session: Session = Depends(get_session)):
 
 
 @router.put("/{equipment_id}", response_model=EquipmentPublic)
-def update_equipment(equipment_id: int, updated_equipment: EquipmentBase, session: Session = Depends(get_session)):
+def update_equipment(equipment_id: int, updated_equipment: EquipmentBase, session: Session = Depends(get_session), current_user: User = Depends(verify_token)):
     equipment = session.get(Equipment, equipment_id)
     if not equipment:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Equipment not found")
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Operation not permitted")
     for key, value in updated_equipment.model_dump().items():
         setattr(equipment, key, value)
     session.commit()
@@ -48,10 +54,12 @@ def update_equipment(equipment_id: int, updated_equipment: EquipmentBase, sessio
 
 
 @router.patch("/{equipment_id}", response_model=EquipmentPublic)
-def partial_update_equipment(equipment_id: int, equipment_update: EquipmentPartialUpdate, session: Session = Depends(get_session)):
+def partial_update_equipment(equipment_id: int, equipment_update: EquipmentPartialUpdate, session: Session = Depends(get_session), current_user: User = Depends(verify_token)):
     equipment = session.get(Equipment, equipment_id)
     if not equipment:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Equipment not found")
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Operation not permitted")
     update_data = equipment_update.model_dump(exclude_unset=True)
     for key, value in update_data.items():
         setattr(equipment, key, value)
@@ -60,9 +68,11 @@ def partial_update_equipment(equipment_id: int, equipment_update: EquipmentParti
     return equipment
 
 @router.delete("/{equipment_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_equipment(equipment_id: int, session: Session = Depends(get_session)):
+def delete_equipment(equipment_id: int, session: Session = Depends(get_session), current_user: User = Depends(verify_token)):
     equipment = session.get(Equipment, equipment_id)
     if not equipment:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Equipment not found")
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Operation not permitted")
     session.delete(equipment)
     session.commit()
